move check_matplotlib/pandas_support outside on utils.__init__
