+++ sklearn/datasets/_openml.py
@@ -15,10 +15,8 @@ from warnings import warn
 
 import numpy as np
 
-from ..utils import (
-    Bunch,
-    check_pandas_support,  # noqa  # noqa
-)
+from ..utils import Bunch
+from ..utils._optional_dependencies import check_pandas_support  # noqa
 from ..utils._param_validation import (
     Integral,
     Interval,
