+++ sklearn/metrics/_plot/regression.py
@@ -2,7 +2,8 @@ import numbers
 
 import numpy as np
 
-from ...utils import _safe_indexing, check_matplotlib_support, check_random_state
+from ...utils import _safe_indexing, check_random_state
+from ...utils._optional_dependencies import check_matplotlib_support
 
 
 class PredictionErrorDisplay:
